feat(project): refactor project structure
