+++ src/blockchain/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_TOKEN_ID =
+  'wSHV2S4qX9jFsLjQo8r1BsMLH2ZRKsZx6EJd1sbozGPieEC4Jf' as const;
