+++ src/context.ts
@@ -1,5 +1,6 @@
 import { Span } from '@opentelemetry/tracing';
-import { ArchiveNodeAdapter, DatabaseAdapter } from './db';
+import { ArchiveNodeAdapter } from 'src/db/archive-node-adapter/archive-node-adapter';
+import { DatabaseAdapter } from 'src/db/archive-node-adapter/archive-node-adapter.interface';
 
 export interface GraphQLContext {
   db_client: DatabaseAdapter;
