+++ src/db/archive-node-adapter/archive-node-adapter.interface.ts
@@ -1,0 +1,10 @@
+import type { EventFilterOptionsInput } from 'src/resolvers-types';
+import type { Actions, Events } from 'src/blockchain/types';
+
+export interface DatabaseAdapter {
+  getEvents(input: EventFilterOptionsInput, options?: unknown): Promise<Events>;
+  getActions(
+    input: EventFilterOptionsInput,
+    options?: unknown
+  ): Promise<Actions>;
+}
