+++ src/services/actions-service/actions-service.interface.ts
@@ -1,0 +1,6 @@
+import { Actions } from 'src/blockchain/types';
+import { ActionFilterOptionsInput } from 'src/resolvers-types';
+
+export interface IActionsService {
+  getActions(input: ActionFilterOptionsInput): Promise<Actions>;
+}
