+++ src/services/actions-service/actions-service.ts
@@ -1,30 +1,32 @@
 import type postgres from 'postgres';
+
 import {
   BlockStatusFilter,
   BlocksWithTransactionsMap,
-  DEFAULT_TOKEN_ID,
   FieldElementIdWithValueMap,
   Action,
   Actions,
-} from '../../models/types';
-import { ActionFilterOptionsInput } from 'src/resolvers-types';
-import { TracingService } from 'src/tracing/tracing';
-import { getActionsQuery } from './queries';
+} from 'src/blockchain/types';
+import type { ActionFilterOptionsInput } from 'src/resolvers-types';
+import type { ITracingService } from 'src/services/tracing-service/tracing-service.interface';
+import { DEFAULT_TOKEN_ID } from 'src/blockchain/constants';
+import { createBlockInfo } from 'src/blockchain/utils';
+import { getActionsQuery } from 'src/db/sql/events-actions/queries';
 import {
   partitionBlocks,
   getElementIdFieldValues,
   mapActionOrEvent,
   removeRedundantEmittedFields,
   sortAndFilterBlocks,
-} from './utils';
-import { createBlockInfo } from '../../models/utils';
+} from 'src/services/utils/utils';
+import { IActionsService } from './actions-service.interface';
 
 export { ActionsService };
 
-class ActionsService {
+class ActionsService implements IActionsService {
   constructor(
     private client: postgres.Sql,
-    private tracingService: TracingService
+    private tracingService: ITracingService
   ) {
     this.client = client;
     this.tracingService = tracingService;
