+++ src/services/events-service/events-service.interface.ts
@@ -1,0 +1,6 @@
+import { Events } from 'src/blockchain/types';
+import { EventFilterOptionsInput } from 'src/resolvers-types';
+
+export interface IEventsService {
+  getEvents(input: EventFilterOptionsInput): Promise<Events>;
+}
