+++ src/services/tracing-service/tracing-service.interface.ts
@@ -1,0 +1,4 @@
+export interface ITracingService {
+  startSpan(name: string): void;
+  endSpan(): void;
+}
